# Improve perf of HttpRuleParser.TryStringToDate