Use a span instead of a string so as to avoid potentially allocating a substring, and first try to use "r" directly in order to take advantage of a fast path it has in TryParseExact; this is the same thing SocketsHttpHandler already does.

Using dotnet/runtime main which has some recently added significant improvements for the DateFormats case (otherwise, using .NET 7 the "Old" case here would be about 3x worse than it shows below):

| Method |      Mean |    Error |   StdDev |
|------- |----------:|---------:|---------:|
|    Old | 383.20 ns | 1.897 ns | 1.863 ns |
|    New |  48.93 ns | 0.204 ns | 0.181 ns |

```C#
private string _date = DateTime.UtcNow.ToString("r");

[Benchmark]
public DateTimeOffset Old() => TryStringToDate_Old(_date, out DateTimeOffset result) ? result : default;

[Benchmark]
public DateTimeOffset New() => TryStringToDate_New(_date, out DateTimeOffset result) ? result : default;

internal static bool TryStringToDate_Old(StringSegment input, out DateTimeOffset result) =>
    DateTimeOffset.TryParseExact(input.ToString(), DateFormats, DateTimeFormatInfo.InvariantInfo,
        DateTimeStyles.AllowWhiteSpaces | DateTimeStyles.AssumeUniversal, out result);

internal static bool TryStringToDate_New(StringSegment input, out DateTimeOffset result)
{
    ReadOnlySpan<char> inputSpan = input.AsSpan();
    DateTimeFormatInfo invariant = DateTimeFormatInfo.InvariantInfo;
    return
        DateTimeOffset.TryParseExact(inputSpan, "r", invariant, DateTimeStyles.None, out result) ||
        DateTimeOffset.TryParseExact(inputSpan, DateFormats, invariant, DateTimeStyles.AllowWhiteSpaces | DateTimeStyles.AssumeUniversal, out result);
}

private static readonly string[] DateFormats = new string[]
{
    // "r", // RFC 1123, required output format but too strict for input
    "ddd, d MMM yyyy H:m:s 'GMT'", // RFC 1123 (r, except it allows both 1 and 01 for date and time)
    "ddd, d MMM yyyy H:m:s", // RFC 1123, no zone - assume GMT
    "d MMM yyyy H:m:s 'GMT'", // RFC 1123, no day-of-week
    "d MMM yyyy H:m:s", // RFC 1123, no day-of-week, no zone
    "ddd, d MMM yy H:m:s 'GMT'", // RFC 1123, short year
    "ddd, d MMM yy H:m:s", // RFC 1123, short year, no zone
    "d MMM yy H:m:s 'GMT'", // RFC 1123, no day-of-week, short year
    "d MMM yy H:m:s", // RFC 1123, no day-of-week, short year, no zone

    "dddd, d'-'MMM'-'yy H:m:s 'GMT'", // RFC 850, short year
    "dddd, d'-'MMM'-'yy H:m:s", // RFC 850 no zone
    "ddd, d'-'MMM'-'yyyy H:m:s 'GMT'", // RFC 850, long year
    "ddd MMM d H:m:s yyyy", // ANSI C's asctime() format

    "ddd, d MMM yyyy H:m:s zzz", // RFC 5322
    "ddd, d MMM yyyy H:m:s", // RFC 5322 no zone
    "d MMM yyyy H:m:s zzz", // RFC 5322 no day-of-week
    "d MMM yyyy H:m:s", // RFC 5322 no day-of-week, no zone
};
```